finished part 1 of day 7

+++ 2024/day_7/part_1/mathe.h
@@ -1,7 +1,8 @@
 
+#include <stdint.h>
+
 #ifndef MATHE_H
 #define MATHE_H
-
 
 
 /**
@@ -9,7 +10,7 @@
  * \param line string to evaluate
  * \return the value of the math in the string, or 0 if it is not possible.
  */
-int m_evaluate_line(char* line);
+unsigned long long m_evaluate_line(char* line);
